+++ developer/src/kmcmplib/src/CompileKeyboardBuffer.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "compfile.h"
+
+bool CompileKeyboardBuffer(KMX_BYTE* infile, int sz, PFILE_KEYBOARD fk);
