+++ developer/src/kmcmplib/src/meson.build
@@ -37,8 +37,10 @@ icuuc_dep = icu.get_variable('icuuc_dep')
 lib = library('kmcmplib',
   'CasedKeys.cpp',
   'CharToKeyConversion.cpp',
+  'CompileKeyboardBuffer.cpp',
   'Compiler.cpp',
   'CompilerInterfaces.cpp',
+  'CompilerInterfacesWasm.cpp',
   'DeprecationChecks.cpp',
   'Edition.cpp',
   'NamedCodeConstants.cpp',
